+++ src/storage/index.ts
@@ -4,5 +4,6 @@ export { getTask } from "./get-task.js";
 export { findHighestTaskId, readHighWaterMark, writeHighWaterMark } from "./hwm.js";
 export { listTasks } from "./list-tasks.js";
 export { ensureTaskListLockFile, ensureTasksDir, LOCK_OPTIONS } from "./locking.js";
+export { resetTaskList } from "./reset.js";
 export { ENV_SNAPSHOT_AT_LOAD, getTaskListId, isInheritedTaskListId, PI_TASK_LIST_ID_ENV } from "./task-list-id.js";
 export { type UpdateTaskResult, updateTask } from "./update-task.js";